move stream row to table object file
