+++ airbyte-webapp-e2e-tests/cypress/pages/connection/createConnectionPageObject.ts
@@ -11,16 +11,6 @@ const connectorHeaderGroupIcon = (connectorType: ConnectorType) =>
   `span[data-testid='connector-header-group-icon-container-${connectorType}']`;
 const catalogTreeTableHeader = `div[data-testid='catalog-tree-table-header']`;
 const catalogTreeTableBody = `div[data-testid='catalog-tree-table-body']`;
-const streamTableRow = (namespace: string, streamName: string) =>
-  `div[data-testid='catalog-tree-table-row-${namespace}-${streamName}']`;
-const streamPanel = (namespace: string, streamName: string) =>
-  `div[data-testid='stream-details-panel-${namespace}-${streamName}']`;
-const streamPanelCloseButton = `button[data-testid='stream-panel-close-button']`;
-const streamSyncSwitch = `label[data-testid='sync-switch']`;
-const sourceStreamNameCell = `div[data-testid='source-stream-name-cell']`;
-const destinationStreamNameCell = `div[data-testid='destination-stream-name-cell']`;
-const sourceNamespaceCell = `div[data-testid='source-namespace-cell']`;
-const destinationNamespaceCell = `div[data-testid='destination-namespace-cell']`;
 
 export const selectExistingConnectorFromDropdown = (connectorName: string) =>
   cy
@@ -78,51 +68,3 @@ export const scrollTableToStream = (streamName: string) => {
 
 export const isStreamTableRowVisible = (streamName: string) =>
   cy.get(catalogTreeTableBody).contains(streamName).should("be.visible");
-
-export const getStreamUtilityFunctions = (namespace: string, streamName: string) => {
-  const stream = streamTableRow(namespace, streamName);
-  const panel = streamPanel(namespace, streamName);
-
-  const isStreamSyncEnabled = (expectedValue: boolean) =>
-    cy.get(stream).within(() => {
-      cy.get(streamSyncSwitch)
-        .get("input")
-        .should(`${expectedValue ? "" : "not."}be.checked`);
-    });
-
-  const toggleStreamSync = () =>
-    cy.get(stream).within(() => {
-      cy.get(streamSyncSwitch).click();
-    });
-
-  const isStreamRowHasRemovedStyle = (expectedValue: boolean) =>
-    cy
-      .get(stream)
-      .invoke("attr", "class")
-      .should(`${expectedValue ? "" : "not."}match`, /removed/);
-
-  const checkSourceNamespace = () => cy.get(stream).within(() => cy.get(sourceNamespaceCell).contains(namespace));
-  const checkSourceStreamName = () => cy.get(stream).within(() => cy.get(sourceStreamNameCell).contains(streamName));
-
-  const checkDestinationNamespace = (expectedValue: string) =>
-    cy.get(stream).within(() => cy.get(destinationNamespaceCell).contains(expectedValue));
-  const checkDestinationStreamName = (expectedValue: string) =>
-    cy.get(stream).within(() => cy.get(destinationStreamNameCell).contains(expectedValue));
-
-  const openStreamPanel = () => cy.get(stream).within(() => cy.get(destinationNamespaceCell).click());
-  const closeStreamPanel = () => cy.get(panel).within(() => cy.get(streamPanelCloseButton).click());
-  const isStreamPanelVisible = (expectedValue: boolean) => cy.get(panel).should(`${expectedValue ? "" : "not."}exist`);
-
-  return {
-    isStreamSyncEnabled,
-    toggleStreamSync,
-    isStreamRowHasRemovedStyle,
-    checkSourceNamespace,
-    checkSourceStreamName,
-    checkDestinationNamespace,
-    checkDestinationStreamName,
-    openStreamPanel,
-    closeStreamPanel,
-    isStreamPanelVisible,
-  };
-};
